[development/allocators] finish all _abspath implementations and runtime bugs that came from it. also fixing some compiler warnings where allocators are using 32bit integers instead of 64bit integers unit tests can finish running now

+++ source/1_engine/rex_engine/include/public/rex_engine/memory/allocators/block_allocator.h
@@ -21,7 +21,7 @@ namespace rex
 	class BlockAllocator
 	{
 	public:
-		using size_type = s32;
+		using size_type = s64;
 		using pointer = void*;
 
 		BlockAllocator(size_type size, size_type blockSize, BackendAllocator alloc = BackendAllocator())
