+++ source/1_engine/rex_engine/include/public/rex_engine/memory/allocators/buddy_allocator.h
@@ -21,7 +21,7 @@ namespace rex
 	class BuddyAllocator
 	{
 	public:
-		using size_type = s32;
+		using size_type = s64;
 		using pointer = void*;
 
 		explicit BuddyAllocator(size_type size, BackendAllocator alloc = BackendAllocator())
